requirements.txt
discord.py[speed]
python-dotenv
aiosqlite # for now 
jishaku